requirements.txt
streamlit>=1.32.0
requests>=2.31.0
python-dotenv>=1.0.0
mistralai>=0.0.7
chromadb>=0.4.22
numpy>=1.24.0